requirements.txt
streamlit>=1.37
inference-sdk>=0.9
python-dotenv>=1.0
Pillow>=10.0
